new file:   Code/MUIT2013/MUIT2013.Business/AttributeDefinitionService.cs 	deleted:    Code/MUIT2013/MUIT2013.Business/ColumnDefinitionService.cs 	new file:   Code/MUIT2013/MUIT2013.Business/DataMiningService.cs 	modified:   Code/MUIT2013/MUIT2013.Business/DataService.cs 	new file:   Code/MUIT2013/MUIT2013.Business/HandlerTrackerService.cs 	modified:   Code/MUIT2013/MUIT2013.Business/MUIT2013.Business.csproj 	modified:   Code/MUIT2013/MUIT2013.Business/packages.config 	modified:   Code/MUIT2013/MUIT2013.Data/MUIT2013.Data.csproj 	renamed:    Code/MUIT2013/MUIT2013.Data/Migration/003_CreateColumnDefinitionsTable.cs -> Code/MUIT2013/MUIT2013.Data/Migration/003_CreateAttributeDefinitionsTable.cs 	modified:   Code/MUIT2013/MUIT2013.Data/Migration/004_CreateMapRulesTable.cs 	new file:   Code/MUIT2013/MUIT2013.Data/Migration/005_CreateHandlerTrackersTable.cs 	renamed:    Code/MUIT2013/MUIT2013.Data/Models/ColumnDefinition.cs -> Code/MUIT2013/MUIT2013.Data/Models/AttributeDefinition.cs 	new file:   Code/MUIT2013/MUIT2013.Data/Models/HandlerTracker.cs 	modified:   Code/MUIT2013/MUIT2013.Data/Models/MapRule.cs 	renamed:    Code/MUIT2013/MUIT2013.Data/Repository/ColumnDefinitionRepository.cs -> Code/MUIT2013/MUIT2013.Data/Repository/AttributeDefinitionRepository.cs 	modified:   Code/MUIT2013/MUIT2013.Data/Repository/DataRepository.cs 	new file:   Code/MUIT2013/MUIT2013.Data/Repository/HandlerTrackerRepository.cs 	new file:   Code/MUIT2013/MUIT2013.DataMining.Models/DecisionTable.cs 	modified:   Code/MUIT2013/MUIT2013.DataMining.Reducts/MUIT2013.DataMining.Reducts.csproj 	new file:   Code/MUIT2013/MUIT2013.DataMining.Reducts/packages.config 	new file:   Code/MUIT2013/MUIT2013.DataMining.Reducts/x86/SQLite.Interop.dll 	modified:   Code/MUIT2013/MUIT2013.DataMining/AttributeRule/AddNumericRule.cs 	modified:   Code/MUIT2013/MUIT2013.DataMining/AttributeRule/INumericRule.cs 	renamed:    Code/MUIT2013/MUIT2013.DataMining/AttributeRule/IGeneralRule.cs -> Code/MUIT2013/MUIT2013.DataMining/AttributeRule/IRule.cs 	modified:   Code/MUIT2013/MUIT2013.DataMining/AttributeRule/IStringRule.cs 	modified:   Code/MUIT2013/MUIT2013.DataMining/AttributeRule/MultiplyNumericRule.cs 	modified:   Code/MUIT2013/MUIT2013.DataMining/AttributeRule/RangeNumericRule.cs 	modified:   Code/MUIT2013/MUIT2013.DataMining/AttributeRule/RangeStringRule.cs 	modified:   Code/MUIT2013/MUIT2013.DataMining/AttributeRule/RegexStringRule.cs 	modified:   Code/MUIT2013/MUIT2013.DataMining/AttributeRule/SingleNumericRule.cs 	modified:   Code/MUIT2013/MUIT2013.DataMining/AttributeRule/SingleStringRule.cs 	modified:   Code/MUIT2013/MUIT2013.DataMining/MUIT2013.DataMining.csproj 	modified:   Code/MUIT2013/MUIT2013.DataMining/packages.config 	modified:   Code/MUIT2013/MUIT2013.Presentation/Forms/ConfigurationForm.Designer.cs 	modified:   Code/MUIT2013/MUIT2013.Presentation/Forms/ConfigurationForm.cs 	modified:   Code/MUIT2013/MUIT2013.Presentation/Forms/ConfigurationForm.resx 	modified:   Code/MUIT2013/MUIT2013.Presentation/Forms/FormBase.cs 	new file:   Code/MUIT2013/MUIT2013.Presentation/Forms/HandlerTrackerForm.Designer.cs 	new file:   Code/MUIT2013/MUIT2013.Presentation/Forms/HandlerTrackerForm.cs 	new file:   Code/MUIT2013/MUIT2013.Presentation/Forms/HandlerTrackerForm.resx 	modified:   Code/MUIT2013/MUIT2013.Presentation/Forms/ImportForm.cs 	modified:   Code/MUIT2013/MUIT2013.Presentation/Forms/ViewDataForm.cs 	modified:   Code/MUIT2013/MUIT2013.Presentation/MUIT2013.Presentation.csproj 	modified:   Code/MUIT2013/MUIT2013.Presentation/MainForm.Designer.cs 	modified:   Code/MUIT2013/MUIT2013.Presentation/MainForm.cs 	renamed:    Code/MUIT2013/MUIT2013.Presentation/Shared/ColumnDefinitionViewFactory.cs -> Code/MUIT2013/MUIT2013.Presentation/Shared/AttributeDefinitionViewFactory.cs 	new file:   Code/MUIT2013/MUIT2013.Presentation/Shared/Events/HandlerTrackerEvent.cs 	renamed:    Code/MUIT2013/MUIT2013.Presentation/Shared/ViewData/ColumnDefinitionView.cs -> Code/MUIT2013/MUIT2013.Presentation/Shared/ViewData/AttributeDefinitionView.cs 	modified:   Code/MUIT2013/MUIT2013.Presentation/Shared/ViewData/NumericRuleColumnDefinitionView.cs 	modified:   Code/MUIT2013/MUIT2013.Presentation/Shared/ViewData/StringRuleColumnDefinitionView.cs 	modified:   Code/MUIT2013/MUIT2013.Tests/MUIT2013.Tests.csproj 	modified:   Code/MUIT2013/MUIT2013.Tests/packages.config 	modified:   Code/MUIT2013/MUIT2013.Utils/MUIT2013.Utils.csproj 	modified:   Code/MUIT2013/MUIT2013.Utils/packages.config 	modified:   Code/MUIT2013/MUIT2013.sln

+++ Code/MUIT2013/MUIT2013.DataMining/AttributeRule/AddNumericRule.cs
@@ -9,13 +9,13 @@ namespace MUIT2013.DataMining.AttributeRule
 {
     public class AddNumericRule : INumericRule
     {
-        public float Add { get; set; }
+        public double Add { get; set; }
         public AddNumericRule()
         {
 
         }
 
-        public float? Apply(float source)
+        public double? Apply(double source)
         {
             return source + Add;
         }
